feat: support watchlib
