+++ packages/rax-plugin-component/package.json
@@ -1,6 +1,6 @@
 {
   "name": "rax-plugin-component",
-  "version": "0.1.9-0",
+  "version": "0.1.9-1",
   "description": "rax component base plugins",
   "license": "BSD-3-Clause",
   "main": "src/index.js",
