+++ packages/rax-plugin-component/src/config/user/default.config.js
@@ -2,6 +2,8 @@ const address = require('address');
 
 module.exports = {
   outputDir: 'lib',
+  devOutputDir: 'lib',
+  devWatchLib: false,
   distDir: 'build',
   publicPath: '/',
   devPublicPath: '/',
