--- packages/rax-plugin-component/src/config/user/keys/outputDir.js
@@ -1,5 +1,0 @@
-const path = require('path');
-
-module.exports = (config, context, value) => {
-  // do nothing for lib Dir
-};
