+++ packages/rax-plugin-component/src/config/user/setConfig.js
@@ -4,7 +4,7 @@ const defaultConfig = require('./default.config');
 module.exports = (config, context, target) => {
   const { userConfig } = context;
 
-  const supportConfig = _.pick(userConfig, Object.keys(defaultConfig));
+  const supportConfig = _.pick(userConfig, Object.keys(_.omit(defaultConfig, ['devWatchLib', 'devOutputDir', 'outputDir'])));
 
   _.forEach(supportConfig, (value, key) => {
     const setKey = require(`./keys/${key}`);
