+++ packages/rax-plugin-component/src/gulpConfig/index.js
@@ -1,0 +1,7 @@
+const registerTasks = require('./registerTasks');
+const filePatterns = require('./filePatterns');
+
+module.exports = {
+  registerTasks,
+  filePatterns
+};
