+++ packages/rax-plugin-component/src/index.js
@@ -5,7 +5,9 @@ const defaultUserConfig = require('./config/user/default.config');
 const dev = require('./dev');
 const build = require('./build');
 
-module.exports = (api, options = {}) => {
+const gulpConfig = require('./gulpConfig');
+
+const PluginComponent = (api, options = {}) => {
   if (!(options.targets && options.targets.length)) {
     console.error(chalk.red('rax-plugin-component need to set targets, e.g. ["rax-plugin-component", targets: ["web", "weex"]]'));
     console.log();
@@ -23,3 +25,7 @@ module.exports = (api, options = {}) => {
     build(api, options);
   }
 };
+
+PluginComponent.gulpConfig = gulpConfig;
+
+module.exports = PluginComponent;
